+++ src/output/saveFileManager/SaveFileManager.cpp
@@ -118,6 +118,13 @@ void SaveFileManager::OnSaveFileCreated()
     ClearEventFlag(static_cast<EventFlags>(FLAG_E102_MR_ENTRANCE_WINDY));
     ClearEventFlag(static_cast<EventFlags>(FLAG_E102_MR_WINDYSTONE));
 
+    //Angel Island Shrine door
+    ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_MR_ISLANDDOOR));
+    ClearEventFlag(static_cast<EventFlags>(FLAG_KNUCKLES_MR_ISLANDDOOR));
+    ClearEventFlag(static_cast<EventFlags>(FLAG_E102_MR_ISLANDDOOR));
+
+
+    
     //Ice Cap Stone
     ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_SS_ICESTONE));
     ClearEventFlag(static_cast<EventFlags>(FLAG_MILES_SS_ICESTONE));