+++ src/output/worldStateManager/WorldStateManager.cpp
@@ -13,7 +13,6 @@ constexpr int WARP_MYSTIC_RUINS = 6;
 constexpr int WARP_EGG_CARRIER_OUTSIDE = 6;
 constexpr int WARP_PAST = 10;
 
-constexpr int COLLISION_CUBE_MYSTIC_RUINS = 42;
 constexpr int SCENE_CHANGE_MYSTIC_RUINS = 33;
 constexpr int RED_MOUNTAIN_DOOR_MYSTIC_RUINS = 15;
 constexpr int LONG_LADDER_MYSTIC_RUINS = 59;
@@ -114,15 +113,6 @@ FunctionHook<void, task*> onCollisionCube(0x4D47E0, [](task* tp) -> void
         if (tp->twp->pos.x < 2 && tp->twp->pos.x > 0
             && tp->twp->pos.y < -1713 && tp->twp->pos.y > -1716
             && tp->twp->pos.z < 2769 && tp->twp->pos.z > 2765)
-            FreeTask(tp);
-    }
-    else if (levelact(CurrentLevel, CurrentAct) == LevelAndActIDs_MysticRuins1
-        && worldStateManagerPtr->unlockStatus.keyDynamite)
-    {
-        //We find the cube collision that we created for the dynamite and delete it
-        if (tp->twp->pos.x < -394 && tp->twp->pos.x > -392
-            && tp->twp->pos.y < 121 && tp->twp->pos.y > 118
-            && tp->twp->pos.z < 891 && tp->twp->pos.z > 889)
             FreeTask(tp);
     }
 
@@ -413,6 +403,7 @@ void WorldStateManager::SetStartingArea()
         break;
     case AngelIsland:
         SetLevelAndAct(LevelIDs_MysticRuins, 1);
+        SetEntranceNumber(1);
         break;
     case NoStatingArea:
         SetLevelAndAct(LevelIDs_StationSquare, 3);
@@ -534,9 +525,6 @@ const SETEntry FINAL_EGG_SPRING = CreateSetEntry(1, {-52.21f, -3240.81f, -190.0f
 const SETEntry SEWERS_SPRING = CreateSetEntry(1, {505, -89, 635},
                                               {0, 0, 0}, {0.3f, 0, 51});
 
-
-const SETEntry COLLISION_CUBE_MR = CreateSetEntry(COLLISION_CUBE_MYSTIC_RUINS, {-393.62f, 120, 890.06f},
-                                                  {0xFEFF, 0x4BF1, 0xFD6A}, {60, 80, 10});
 const SETEntry RED_MOUNTAIN_SCENE_CHANGE_MR = CreateSetEntry(SCENE_CHANGE_MYSTIC_RUINS, {-2100, -304, 1650},
                                                              {0, 0, 0}, {40, 50, 0});
 
@@ -630,13 +618,6 @@ FunctionHook<void> onCountSetItemsMaybe(0x0046BD20, []()-> void
 
     AddSetToLevel(FINAL_EGG_SPRING, LevelAndActIDs_FinalEgg3, Characters_Sonic);
     AddSetToLevel(SEWERS_SPRING, LevelAndActIDs_StationSquare3, Characters_Sonic);
-
-    AddSetToLevel(COLLISION_CUBE_MR, LevelAndActIDs_MysticRuins1, Characters_Sonic);
-    AddSetToLevel(COLLISION_CUBE_MR, LevelAndActIDs_MysticRuins1, Characters_Tails);
-    AddSetToLevel(COLLISION_CUBE_MR, LevelAndActIDs_MysticRuins1, Characters_Knuckles);
-    AddSetToLevel(COLLISION_CUBE_MR, LevelAndActIDs_MysticRuins1, Characters_Amy);
-    AddSetToLevel(COLLISION_CUBE_MR, LevelAndActIDs_MysticRuins1, Characters_Big);
-    AddSetToLevel(COLLISION_CUBE_MR, LevelAndActIDs_MysticRuins1, Characters_Gamma);
 
     AddSetToLevel(RED_MOUNTAIN_SCENE_CHANGE_MR, LevelAndActIDs_MysticRuins2, Characters_Tails);
     AddSetToLevel(RED_MOUNTAIN_DOOR_MR, LevelAndActIDs_MysticRuins2, Characters_Tails);