+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
   locations of the life capsules and field emblems. (Credit to DementisXYZ).
 - Re-enabled the Egg Carrier transformation cutscene to help players know that the switch worked.
     - The cutscene is unskippable, so an option in the mod configuration was added to disable it.
+- When starting on Angel Island, the player will spawn near the Red Mountain entrance.
+    - The door to the cave will be closed until you get the dynamite.
 
 ### Changed